refactor: move _git to utils.py
